Set DISPLAY to :0 when not set
WSL sessions don't inherit the DISPLAY environment variable, so we have
to set it defensively.

This shouldn't be a problem on actual Linux desktops because they'll
have DISPLAY set correctly already. And Linux headless servers will
either not have a display (in which case we will default to :0 and
continue to not work) or SSH will set it for the client and we won't
change it.