Upgrade Libertinus and Iosevka fonts and enable Git LFS
Iosevka has moved to a TTC collection, which is very nice. For
Libertinus I've included the entire bundle contents for simplicity,
rather than picking out the font files.

Because of the TTC collections for Iosevka, I've had to enable Git LFS
for this repository. I should possibly have done that as a separate
commit, but adding the filters also replaced the file contents, so it
would be too much bother to try to rewrite that history now.