Configure helm to ignore .yarn directory for projectile
Change helm to use a hybrid indexing approach, so that it will apply the filters
in projectile-globally-ignored-directories (and similar variables). By default
it uses alien indexing, which uses purely find -- no filtering of the results.
This may be slower for very large projects, but I think it's worth it.

Note: This is only for projectile (SPC p f) -- i.e., "project files." Projectile
already ignores files in .gitignore (or more likely, uses `git ls-files`), but
we've added files in .yarn to git for yarn's "zero install" feature.